requirements.txt
streamlit
langchain
langchain-community
langchain-chroma>=0.1.2
langchain-huggingface
sentence-transformers
langchain-groq
langchain-text-splitters
huggingface-hub
sentence-transformers
python-dotenv

